+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 ## [Unreleased]
 ### Added
 - this changelog
+- support for module selection params (i.e. `{ commonjs: true }`) via rule options
 
 ## [0.1.0] - 2016-04-05
 Born.